Create GetUserInfo
This project is called GetUserInfo. in this project we were tasked with making a code that would be able to ask the user to input their first and last name, and age. Then we would calculate how old they would be in Ten plus years and ask if we can call them by their first name.